requirements.txt
Django==1.7
djangorestframework==3.1.1
Pillow==5.0.0